Separate actuator info feature
