+++ sip-core/src/main/resources/META-INF/spring.factories
@@ -3,4 +3,5 @@ de.ikor.sip.foundation.core.translate.TranslationsAutoConfig,\
 de.ikor.sip.foundation.core.trace.TraceAutoConfig,\
 de.ikor.sip.foundation.core.proxies.ProxyAutoConfig,\
 de.ikor.sip.foundation.core.actuator.routes.ActuatorRoutesAutoConfig,\
-de.ikor.sip.foundation.core.actuator.health.ActuatorHealthAutoConfig
+de.ikor.sip.foundation.core.actuator.health.ActuatorHealthAutoConfig,\
+de.ikor.sip.foundation.core.actuator.info.ActuatorInfoAutoConfig
