+++ sip-core/src/main/resources/sip-core-default-config.yaml
@@ -18,6 +18,8 @@ sip.core:
       showExchangeId: true
     traceType: "*"
   actuator:
+    info:
+      enabled: true
     routes:
       enabled: true
     health:
