+++ sip-core/src/test/java/de/ikor/sip/foundation/core/actuator/info/AdapterInfoContributorTest.java
@@ -1,4 +1,4 @@
-package de.ikor.sip.foundation.core.actuator.routes;
+package de.ikor.sip.foundation.core.actuator.info;
 
 import static org.assertj.core.api.Assertions.assertThat;
